+++ ft_putstr.c
@@ -6,15 +6,15 @@
 /*   By: dphyliss <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/23 15:58:44 by dphyliss          #+#    #+#             */
-/*   Updated: 2019/04/23 16:23:27 by dphyliss         ###   ########.fr       */
+/*   Updated: 2019/04/24 12:10:30 by dphyliss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putstr(char const *s)
+void		ft_putstr(char const *s)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s)
