+++ ft_strclr.c
@@ -6,15 +6,15 @@
 /*   By: dphyliss <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/20 17:53:36 by dphyliss          #+#    #+#             */
-/*   Updated: 2019/04/20 17:58:02 by dphyliss         ###   ########.fr       */
+/*   Updated: 2019/04/24 15:04:28 by dphyliss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_strclr(char *s)
+void		ft_strclr(char *s)
 {
-	int	i;
+	size_t	i;
 
 	if (s)
 	{
