+++ ft_strequ.c
@@ -6,15 +6,15 @@
 /*   By: dphyliss <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/22 13:56:53 by dphyliss          #+#    #+#             */
-/*   Updated: 2019/04/22 14:05:44 by dphyliss         ###   ########.fr       */
+/*   Updated: 2019/04/24 15:01:12 by dphyliss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strequ(char const *s1, char const *s2)
+int			ft_strequ(char const *s1, char const *s2)
 {
-	int	i;
+	size_t	i;
 
 	if (s1 && s2)
 	{
